feat: refactor navigation component, extract user menu and sub-navigation into separate components
